Add new database renaming option.
This option is used to add values from the database to the new filename.
Right now only a few values are implemented, but the parser is easy to extend,
so anybody who can think of new values can add them :-)

The CameraUI will not have the database renaming option, because the images are
not yet added to the database (sure).

svn path=/trunk/extragear/graphics/digikam/; revision=1129389