First implementation of "directory creation" with AdvancedRename. Right now this only works in the AlbumUI, not BQM or CameraUI. I guess we need something common for renaming, like a thread running in the background that is used by all the components, not only the AdvancedRenameDialog like it is done now.
To test it, use some test images, and rename them by using a "/" character,
for example:

"[date]/[file]_####"

or something like that.

It can also be nested:

"[ext]/[date]/[file]_####"

But it is not really stable and some directory combinations don't work.
So testing should really be done with some test images!!!

Andi

svn path=/trunk/extragear/graphics/digikam/; revision=1137326