Port FilmGrainFilter to RandomNumberGenerator.
Please check these changes:

1)
    d->generator.seed(1); // noise will always be the same

As before, the generator is initialized with a constant seed.
We could just as well initialize with a random seed.

2)

-    return ((double)(qrand() % (int)range) - range/2.0) ;
+    return d->generator.number(- range/2, range/2);

Before, the randomizeUniform returned random values in steps
of 1 (cast to int), e.g., if range was 5, (-2.5, -1.5, -0.5, 0.5, 1.5, 2.5).
The method name promises a uniform distribution (any value in [-2.5, 2.5]),
and that's what it's now doing.


svn path=/branches/extragear/graphics/digikam/; revision=1193032