Unleash non-destructive editing: The interface is now fully adapted to versioning (includes a first toolbar action cleanup). Yes, Save/Save As is gone. Icons perfectly fitting to our concepts are missing, so I'm using what's available. (I filed a request for icons, but dont know when/if it will be done)
I plan to extend the setup a bit, possibly allowing to switch off
non-destructive completely for those that really want,
and to define the default close action.
(We had planned to default to "save current" when closing the window,
but now I think that per default, the user must be asked)

You'll also see that "Revert to original" kind of changed its nature.
There's a "Load Original", but that's like opening a file.
Then there's a "discard changes", which is effectively like clicking
Undo to the end. Something like "remove the current version, delete
the version file and all its intermediates up to the original" has a lot of
implications, if other files were forked in between.


svn path=/branches/extragear/graphics/digikam/; revision=1200726