Here we go again... :-)
Adapt digiKam coding style guides:
This is a huge patch to finally have a common coding style in the source
code.

This patch adds:

    - curly brackets to all single line statements
    - left pointer / reference operator style (char* bla)
    - indent 4 spaces
    - convert tab -> spaces

I did not change the libs from 3rdparty and the ones from libs/jpegutils.
Code compiles fine under Linux, there shouldn't be code breakage because we
only reformat the code and add missing brackets. If code is broken now
(logic
errors due to missing brackets), it was broken before.

I add this as one single patch so we can easily revert it, if needed.
I think it is really important to have one common style in the code, and
that
we don't mix it.

I will also add some comments to the HACKING file later on.

Gilles,

can you check if all compiles fine under Windows and Mac?

Andi

svn path=/trunk/extragear/graphics/digikam/; revision=1200823