Add KMemoryInfo to access information about system memory.
This class was initially written by Pino for inclusion to kdelibs,
but rejected for various reasons.
The API of this version is changed pretty much to my taste,
the backend is replaced with code copied from libstatgrab, which also
supports various Unixes in addition to Window and Linux.

Note: only tested on Linux so far.

Pretty obviously, this is stuff for kdelibs, but as we saw,
it did not find its way there.

svn path=/branches/extragear/graphics/digikam/core/; revision=1217347