Move Rating filter from statusbar to Tags Filter view from right sidebar. As Rating is considerated as a Labels (like Pick or Color), it need to be moved near Pick/Color Label filters.
Note that i plan to move all filters from status bar to right Sidebar and rename "Tag Filters" as "Filters", and to add
2 buttons on status bar :
"Reset" to clean all active filters
"Settings" (or something like that) to switch on "Filters" panel from right sidebar.

The advantage to use "Filters" panel from right side bar is to be able to host more filters in the future.
Status bar is limited in space to host more settings. There are a lots of whishes about this subject in bugzilla.

CCBUGS: 241847