Remaining glue code to make grouping work.
Current status:
There is basic support for grouping by drag-drop and context menu.
TODO: The visual representation is to be improved - we'll probably need some
custom drawing, as there are no suitable icons. Some visual indication
for items that come from opened groups.
The actions are very context-sensitive, so adding to the main menu
is not straight-forward.

CCMAIL: digikam-devel@kde.org