CollectionManager: fix for collections on network shares
QUrl default constructor expects an unencoded URL. If an encoded
URL is used, the '%' is again encoded and the resulting URL
is unusable. AlbumRootLocation::identifier seems to be already encoded,
therefore use the proper QUrl::fromEncoded() method.

Without this fix, existing collections on network shares are always
marked unavailable and new collections can not be generated.

Signed-off-by: Matthias Welwarsky <matze@welwarsky.de>