Fix thumbnail database with SQLite
For the last few days, the main database would be used as thumbnail database.
The reason is that the backend always read the main fields of the DatabaseParameters,
now prefixed "img". For starting the thumbnail database, the thumbnailParameters()
method is used to retrieve a DatabaseParameters object with the "tmb" parameters
copied to the main fields. This method was a no-op.

The API is not that elegant now but at least it works.