The last commit broke compilation because a "break" was added to the MetadataInfo::Orientation case. The problem is that if no value is matched inside the switch, we do not return a value for the orientation.
But the bigger problem is that before the "break" statement we would have
fallen through to the MetadataInfo::Make case statement, which is totally wrong
in this case.

So please do not ignore compile warnings, they are there for a reason :-)
Always add "break" statements, even in the default handling. If a fall-through
is valid, add a comment to the source code to make the situation clear to other
coders.