The "hide thumbbar in fullscreen mode" setting was not respected in albumUI, only in the image editor. Also the state before entering fullscreen mode was not restored properly, e.g. the toolbar was shown, although the user might have hidden it in the first place.
While fixing the issue, I asked myself: We do we have these two fullscreen mode
settings in the editor window page of the setup dialog? They are used for
albumUI as well (maybe even for lighttable?).

Shouldn't these settings be moved to the "General" page?
Also, an action is missing in the context menu for returning from the
fullscreen mode if the toolbar is hidden. How should a normal user know the
shortcut for this action?
At least my parents wouldn't know it, and I guess they can be considered as
"normal users" ;-)

Andi

CCMAIL:digikam-devel@kde.org