Add movedFrom and movedTo signals to prevent missing file change information
The KInotify::move signal is only emitted if both source and target are being
watched. As we only watch the digikam collections, we miss anything moved from
or to the outside, including moving to the trash.
Instead, listen to the movedFrom and movedTo signals which are emitted without
such preconditions