Enable Color Channels histogram type in Levels tool
This patch enables the "Color Channels" histogram type in the Levels
tool. It is helpful when adjusting luminosity to observe if any of the
color channels clip, which leads to false highlight colors.

The patch also implements a second function:
if Color Channels is selected, Black and White tone pickers work on all
of the color channels simultaneously instead of having to iterate over
R, G and B. It is very helpful for removing the color mask of a scanned
film negative.