Use QScopedArrayPointer here, to avoid memory issues.
We are using a third party library for loading TIFF files. If an exception is thrown or anything else happens so that the loading method is exited in an unusual way, the "delete" statements will never be executed. Since we can not know when the method will exit exactly, a ScopedArrayPointer is the safest way.
In general we should never call "delete" manually if not absolutely necessary, let the compiler / runtime handle this.