Make possible to compile with lcms2
This is the initial phase of the porting, a wrapper has been placed
inbetween lcms and digikam. While for lcms1 it's a simple redirect for
lcms2 instead function are executed, mostly to cast values.

As is digikam loses functionality (and possibly stability) because of
the incomplete/untested code with lcms2, for this reason THE DEFAULT
STILL IS LCMS1, obviously this has to change in the near future.

function that still need porting are:
- cmsErrorAction
- cmsGetProfileICCversion
- cmsTake*