Add missing LCMS version 2 cmake detection script. It's the same than libkdcraw.
Laurent, it must be fine to factorize this script in kdelibs. I know that krita use lcms2 too...
Note there is already a cmake script to detect LCMS version 1 which is very different. Of course,
if this script can be patch to support version 1 and version 2, it will be great.

CCMAIL: montel@kde.org
CCBUGS: 294105