Preserve recently assigned / used tags
This commit contains a dirty hack, because writeSettings() is called every time a tag was added / removed to an image. This is currently needed because the destructor of the AlbumDB object is never called. Setting a breakpoint in its destructor shows that it never reaches the code.

Marcel, is this a problem, besides possibly having a memory leak here?

Andi

CCMAIL:marcel.wiesweg@gmx.de

CCBUG:278345