speed up thumbnail generation maintenance tool.
Use a QSet for subtracting invalid filepaths, this is faster than using the iterator, it::erase and QStringList::contains.
Do not extract the image id, it is never used and querying for it only wastes time.
This will speed up the maintenance tool (in combination with the "scan for images without thumbnail" option)