Revert 1fe432d0828e53c9820b1f27b9a81c8722d7e539 which caused a crash.
MyTile is only deleted by GPSMarkerTiler::tileDelete.
All subclasses of AbstractMarkerTiler have to reimplement tileDelete
to delete their Tile subclasses.
This was done in order not to have any virtual functions
in Tile and its subclasses in order to save memory, since there
can be a lot of tiles in a MarkerTiler.