Fix thumbnails disappearing when columns are too wide.
The idea here is that for landscape images, we adjust the height to
to be as high as the row height as long as the width can stretch enough
because the column is wider than the thumbnail size.
However, digiKam limits the thumbnail size to 256, so we also do that here.

BUG: 319294