Fix copying of DisjointMetadata's data
The title and comment were not copied at all in the manual copy
constructor. Use C++'s automatic copy constructor to be safe also for
future new fields. For this, the Private class needed to be split in a
value container and the actual Private, which has some non-copyable
parts (the mutex)