Apply patch #101823 to make listDateRange() MYSQL 5.7 compliant
MYSQL 5.7 with default configuration introduced stricter query structure constraints.
This patch changes the ORDER BY clause to use Images.album instead of Albums.id, which causes an error since it is not in the SELECT DISTINCT clause.
As can be seen from the INNER JOIN clause Albums.id=Images.album this change does not alter the logic of the query.

BUGS: 370744
FIXED-IN: 5.4.0