I introduced the new special tag Unconfirmed which is automatically created if not present. In detection phase, faces are marked as unknown as usual. In recognition phase, the faces are marked with the recognised face tag, as usual. But additionally, they are connected as autodetected person to the Unconfirmed tag. In this property, I store the id of the person as which the face was recognised and also the face region.
With small adoptions in the image lister all recognised faces are now shown both in the face tag which was recognised (person A) and the Unconfirmed tag (with the suggested tag being person A).

Confirming the person can now be done in both tags. And the confirmed identity is set as expected. Confirming the face also removes the face from unconfirmed tag.
Rejecting the face also removes the face from both the recognised tag and also from the Unconfirmed tag. Rescanning for faces (detection) also removes the connection
to the Unconfirmed tag if unconfirmed results should be thrown away.

BUGS: 336253
FIXED-IN: 5.5.0