New digiKam Video SlideShow tool based on QtAV/FFMpeg encoder.
The tool replace the older Qt4 kipi plugin based on ImageMagick and GStreamer which never
work as expected due to run time dependencies with GStreamer which are a big puzzle to resolve
and don't work under MacOS and Windows.
The tool permit to encode images as a video diaporama.
The tool support many video codecs as X264 or MPEG4 provided by FFMpeg and
is able to store stream as AVI, MKV, or MP4 media.
The tool permit to apply a transition effect between images (same effects than presentation tool non-OpenGL).
The tool permit to export image from current select or a set of albums selection (as HTMLGallery tool).
The tool do not permit to encode an audio soundtrack with the video stream due a bug in QtAV which
have been reported and must be fixed in next QtAV release 1.12.0.
The tool is available currently under digiKam album GUI, but it will be also available into
ImageEditor, LightTable, and Showfoto (as HTMLGallery, Panorama, Calendar, etc... tools.)

Screeshots of tool in action : https://www.flickr.com/photos/digikam/albums/72157684569983835

CCBUGS: 271751
BUGS: 316791
BUGS: 320725
BUGS: 337605
BUGS: 331625
BUGS: 324646
BUGS: 176727
BUGS: 318297
BUGS: 153061
BUGS: 334089
FIXED-IN: 5.6.0