Huge improvement of Media Server here : Serve all image formats supported by digiKam core engine.
The image is not passed as well as an bytes stream as file is read and push on web server, delegating the
mime type analysis and decoding to the renderer.

The idea here is to transcoding the image by extracting the reduced image preview and share it as JPEG.

This will prevent to bloat the network with huge data, reduce time latency with huge image (as panorama),
and prevent crash in client renderer application in case of reduced memory available to host full image data.

By this way we can serve all image formats supported by digiKam, including all RAW file formats.
As i know there is no Upnp/DLNA media server able to share RAW files on the network with this kind of transparency.

In case of non image format to serve, the file is shared as well, as previously. There is no transcoding.

TODO: We need to do something similar with video files, as to transcoding media with a low band of data on the network.
The question is : ffmpeg is able to do it in background (as Kodi do in delegate implementation), but QtAV is able to do it
with current API without to pass directly to ffmpeg ?

CCMAIL: digikam-devel@kde.org