Hide all selected albums when duplicates are removed
I did not dig deeper enough, but my suspicion is that setting the item to hidden messes up with the tree widget
iterator, causing every selected album that is not the first to not hide when all duplicate images are removed.

Switching to a predefined list of selected items resolves the issue.